PIG-1135: skewed join partitioner returns negative partition index  (yinghe via olgan)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/pig/branches/branch-0.6@888868 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/backend/hadoop/executionengine/mapReduceLayer/partitioners/SkewedPartitioner.java
@@ -70,8 +70,8 @@ public class SkewedPartitioner implements Partitioner<PigNullableWritable, Writa
     public int getPartition(PigNullableWritable wrappedKey, Writable value,
             int numPartitions) {
 		// for streaming tables, return the partition index blindly
-		if (wrappedKey instanceof NullablePartitionWritable && ((int)((NullablePartitionWritable)wrappedKey).getPartition()) != -1) {
-			return (int) ((NullablePartitionWritable)wrappedKey).getPartition();
+		if (wrappedKey instanceof NullablePartitionWritable && (((NullablePartitionWritable)wrappedKey).getPartition()) != -1) {
+			return ((NullablePartitionWritable)wrappedKey).getPartition();
 		}
 
 		// for partition table, compute the index based on the sampler output
